requirements-dev.txt
# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0 # Добавлено для отчетов о покрытии
pytest-asyncio==0.21.1
factory-boy==3.3.0
hypothesis==6.92.1
responses==0.24.1
python-slugify==8.0.1 # <-- ДОБАВЛЕНО

# Code Quality & Typing
black==24.4.2
ruff==0.4.1
mypy==1.10.1
django-stubs==5.0.2
djangorestframework-stubs==3.15.0
pre-commit==3.7.1

# Debugging
django-debug-toolbar==6.0.0